Moved ApolloProvider.
