+++ assets/prototype/infrastructure/TopLevelProviders/index.js
@@ -1,5 +1,5 @@
 /* Internal dependencies */
-import ApolloProvider from './Apollo';
+import ApolloProvider from '../../apollo/Apollo';
 
 /**
  * A collection of top level providers that are used by multiple parts of the application.
